+++ src/views/JoinedBroadcastsPage.vue
@@ -5,8 +5,14 @@
     </ion-header>
     <ion-content>
       <NoBroadcasts v-if="!broadcastStore.getJoinedAndInsertedBroadcasts.length" />
-      <BroadcastCard class="my-2" v-for="broadcast in broadcastStore.getJoinedAndInsertedBroadcasts" :broadcast="broadcast" @click-join="onClickJoin(broadcast)">
-        <ion-button fill="solid" expand="full" :style="{margin: '0px'}" class="no-margin" @click="router.push(`/home/joined-broadcasts/${broadcast.id}`)">Chat</ion-button>
+      <BroadcastCard
+        class="my-2"
+        v-for="broadcast in broadcastStore.getJoinedAndInsertedBroadcasts"
+        :broadcast="broadcast"
+        @click-join="onClickJoin(broadcast)">
+        <ion-button fill="solid" expand="full" :style="{ margin: '0px' }" class="no-margin" @click="router.push(`/home/joined-broadcasts/${broadcast.id}`)"
+          >Chat</ion-button
+        >
       </BroadcastCard>
     </ion-content>
   </ion-page>
@@ -50,7 +56,7 @@ onMounted(async () => {
 </script>
 
 <style scoped>
-ion-button{
-  --box-shadow: none
+ion-button {
+  --box-shadow: none;
 }
 </style>
